// frontend/src/sections/dashboard/dashboardCounts.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type InboundShipmentCountsVariables = {
  storeId: string;
};

export type InboundShipmentCountsResult = {
  inboundShipmentCounts: {
  created: {
  today: number;
  thisWeek: number;
};
  notDelivered: number;
};
};

export const InboundShipmentCounts = {
  query: "query inboundShipmentCounts($storeId: String!) {\n  inboundShipmentCounts(storeId: $storeId) {\n    created {\n      today\n      thisWeek\n    }\n    notDelivered\n  }\n}",
} as TypedDocument<InboundShipmentCountsResult, InboundShipmentCountsVariables>;

export type InboundShipmentExternalCountsVariables = {
  storeId: string;
};

export type InboundShipmentExternalCountsResult = {
  inboundShipmentExternalCounts: {
  created: {
  today: number;
  thisWeek: number;
};
  notDelivered: number;
};
};

export const InboundShipmentExternalCounts = {
  query: "query inboundShipmentExternalCounts($storeId: String!) {\n  inboundShipmentExternalCounts(storeId: $storeId) {\n    created {\n      today\n      thisWeek\n    }\n    notDelivered\n  }\n}",
} as TypedDocument<InboundShipmentExternalCountsResult, InboundShipmentExternalCountsVariables>;

export type RequisitionCountsVariables = {
  storeId: string;
};

export type RequisitionCountsResult = {
  requisitionCounts: {
  request: {
  draft: number;
};
  response: {
  new: number;
};
  emergency: {
  new: number;
};
};
};

export const RequisitionCounts = {
  query: "query requisitionCounts($storeId: String!) {\n  requisitionCounts(storeId: $storeId) {\n    request {\n      draft\n    }\n    response {\n      new\n    }\n    emergency {\n      new\n    }\n  }\n}",
} as TypedDocument<RequisitionCountsResult, RequisitionCountsVariables>;

export type OutboundShipmentCountsVariables = {
  storeId: string;
};

export type OutboundShipmentCountsResult = {
  outboundShipmentCounts: {
  notShipped: number;
};
};

export const OutboundShipmentCounts = {
  query: "query outboundShipmentCounts($storeId: String!) {\n  outboundShipmentCounts(storeId: $storeId) {\n    notShipped\n  }\n}",
} as TypedDocument<OutboundShipmentCountsResult, OutboundShipmentCountsVariables>;

export type StockCountsVariables = {
  storeId: string;
  daysTillExpired?: number | null;
};

export type StockCountsResult = {
  stockCounts: {
  expired: number;
  expiringSoon: number;
  expiringInNextThreeMonths: number;
  expiringBetweenThresholds: number;
};
};

export const StockCounts = {
  query: "query stockCounts($storeId: String!, $daysTillExpired: Int) {\n  stockCounts(storeId: $storeId, daysTillExpired: $daysTillExpired) {\n    expired\n    expiringSoon\n    expiringInNextThreeMonths\n    expiringBetweenThresholds\n  }\n}",
} as TypedDocument<StockCountsResult, StockCountsVariables>;

export type ItemCountsVariables = {
  storeId: string;
  lowStockThreshold?: number | null;
  highStockThreshold?: number | null;
};

export type ItemCountsResult = {
  itemCounts: {
  itemCounts: {
  total: number;
  noStock: number;
  lowStock: number;
  highStock: number;
  outOfStockProducts: number;
  productsAtRiskOfBeingOutOfStock: number;
  productsOverstocked: number;
};
};
};

export const ItemCounts = {
  query: "query itemCounts($storeId: String!, $lowStockThreshold: Float, $highStockThreshold: Float) {\n  itemCounts(\n    storeId: $storeId\n    lowStockThreshold: $lowStockThreshold\n    highStockThreshold: $highStockThreshold\n  ) {\n    itemCounts {\n      total\n      noStock\n      lowStock\n      highStock\n      outOfStockProducts\n      productsAtRiskOfBeingOutOfStock\n      productsOverstocked\n    }\n  }\n}",
} as TypedDocument<ItemCountsResult, ItemCountsVariables>;